flex_and_vision/requirements.txt
Flask==0.12.2
gunicorn==19.7.1
google-cloud-vision==0.30.1
google-cloud-storage==1.8.0
google-cloud-datastore==1.6.0
simplejson
